Extract serialized error formatting into its own module
It's only used by reporters, and the other methods in the
format-assert-error module are only used by test workers.
