+++ lib/format-assert-error.js
@@ -4,7 +4,6 @@ const reactTestPlugin = require('@ava/pretty-format/plugins/ReactTestComponent')
 const chalk = require('chalk');
 const diff = require('diff');
 const DiffMatchPatch = require('diff-match-patch');
-const indentString = require('indent-string');
 const globals = require('./globals');
 
 function formatValue(value, options) {
@@ -96,26 +95,3 @@ function formatWithLabel(label, value) {
 	return {label, formatted: formatValue(value)};
 }
 exports.formatWithLabel = formatWithLabel;
-
-function formatSerializedError(error) {
-	if (error.statements.length === 0 && error.values.length === 0) {
-		return null;
-	}
-
-	let result = error.values
-		.map(value => `${value.label}\n\n${indentString(value.formatted, 2).trimRight()}\n`)
-		.join('\n');
-
-	if (error.statements.length > 0) {
-		if (error.values.length > 0) {
-			result += '\n';
-		}
-
-		result += error.statements
-			.map(statement => `${statement[0]}\n${chalk.grey('=>')} ${statement[1]}\n`)
-			.join('\n');
-	}
-
-	return result;
-}
-exports.formatSerializedError = formatSerializedError;
