+++ lib/reporters/mini.js
@@ -9,10 +9,10 @@ const cliTruncate = require('cli-truncate');
 const cross = require('figures').cross;
 const indentString = require('indent-string');
 const ansiEscapes = require('ansi-escapes');
-const formatAssertError = require('../format-assert-error');
 const extractStack = require('../extract-stack');
 const codeExcerpt = require('../code-excerpt');
 const colors = require('../colors');
+const formatSerializedError = require('./format-serialized-error');
 const improperUsageMessages = require('./improper-usage-messages');
 
 class MiniReporter {
@@ -177,7 +177,7 @@ class MiniReporter {
 				}
 
 				if (test.error.avaAssertionError) {
-					const formatted = formatAssertError.formatSerializedError(test.error);
+					const formatted = formatSerializedError(test.error);
 					if (formatted) {
 						status += '\n' + indentString(formatted, 2);
 					}
