+++ lib/reporters/verbose.js
@@ -4,10 +4,10 @@ const prettyMs = require('pretty-ms');
 const figures = require('figures');
 const chalk = require('chalk');
 const plur = require('plur');
-const formatAssertError = require('../format-assert-error');
 const extractStack = require('../extract-stack');
 const codeExcerpt = require('../code-excerpt');
 const colors = require('../colors');
+const formatSerializedError = require('./format-serialized-error');
 const improperUsageMessages = require('./improper-usage-messages');
 
 class VerboseReporter {
@@ -117,7 +117,7 @@ class VerboseReporter {
 				}
 
 				if (test.error.avaAssertionError) {
-					const formatted = formatAssertError.formatSerializedError(test.error);
+					const formatted = formatSerializedError(test.error);
 					if (formatted) {
 						output += '\n' + indentString(formatted, 2);
 					}
